Integrate enhanced modules
