+++ rrc_toolkit.py
@@ -1,8 +1,15 @@
 import argparse
-from ue_mobility_controller import UEMobilityController
+import os
+import sys
+
+ENHANCED_PATH = os.path.join(os.path.dirname(__file__), "srsRAN_5G", "srsRAN_5G")
+if ENHANCED_PATH not in sys.path:
+    sys.path.insert(0, ENHANCED_PATH)
+
+from enhanced_ue_mobility_controller_v2 import main as mobility_main
 from rrc_trace_capture import RRCTraceCapture
-from rrc_trace_analyzer import RRCTraceAnalyzer
-from rrc_trace_visualizer import RRCTraceVisualizer
+from enhanced_rrc_trace_analyzer import main as analyzer_main
+from enhanced_visualization_dashboard import app as dashboard_app
 
 
 def main() -> None:
@@ -17,17 +24,14 @@ def main() -> None:
     args = parser.parse_args()
 
     if args.command == "mobility":
-        controller = UEMobilityController()
-        controller.run_mobility_scenario()
+        mobility_main()
     elif args.command == "capture":
         capture = RRCTraceCapture()
         capture.capture_all_traces()
     elif args.command == "analyze":
-        analyzer = RRCTraceAnalyzer()
-        analyzer.run_analysis()
+        analyzer_main()
     elif args.command == "visualize":
-        visualizer = RRCTraceVisualizer()
-        visualizer.run_visualization()
+        dashboard_app.run_server(debug=False, host="0.0.0.0", port=8050)
     else:
         parser.print_help()
