+++ srsRAN_5G/srsRAN_5G/enhanced_ue_mobility_controller_v2.py
@@ -18,146 +18,12 @@ import threading
 import numpy as np
 import matplotlib.pyplot as plt
 from datetime import datetime
-from abc import ABC, abstractmethod
-
-# --- 基礎類 --- 
-
-class Position:
-    """位置類，表示 UE 在二維平面上的位置"""
-    def __init__(self, x=0, y=0):
-        self.x = x
-        self.y = y
-    
-    def distance_to(self, other):
-        """計算到另一個位置的距離"""
-        return math.sqrt((self.x - other.x)**2 + (self.y - other.y)**2)
-    
-    def __str__(self):
-        return f"({self.x:.2f}, {self.y:.2f})"
-
-# --- 信道模型接口 --- 
-
-class ChannelModel(ABC):
-    """信道模型接口基類"""
-    @abstractmethod
-    def calculate_rsrp(self, gnb, ue_position):
-        """計算 UE 在給定位置相對於特定 gNB 的 RSRP"""
-        pass
-
-    @abstractmethod
-    def calculate_sinr(self, gnb, ue_position, interfering_gnbs):
-        """計算 UE 在給定位置相對於特定 gNB 的 SINR"""
-        pass
-
-    @abstractmethod
-    def is_in_coverage(self, gnb, ue_position, threshold=-110):
-        """判斷 UE 是否在 gNB 的覆蓋範圍內（基於 RSRP 閾值）"""
-        pass
-
-class SimplifiedChannelModel(ChannelModel):
-    """簡化的信道模型實現（基於路徑損耗和隨機衰落）"""
-    def calculate_rsrp(self, gnb, ue_position):
-        """計算 UE 在給定位置的 RSRP
-        使用簡化的路徑損耗模型: RSRP = 發射功率 - 路徑損耗
-        路徑損耗 = 32.4 + 20*log10(頻率) + 20*log10(距離)，頻率單位為 MHz，距離單位為 km
-        """
-        distance_km = gnb.position.distance_to(ue_position) / 1000.0
-        # 避免距離為零導致計算錯誤
-        if distance_km < 0.001:
-            distance_km = 0.001
-        
-        # 計算路徑損耗
-        try:
-            path_loss = 32.4 + 20 * math.log10(gnb.frequency) + 20 * math.log10(distance_km)
-        except ValueError:
-             # 如果 log10 參數無效，返回極小值
-             return -float("inf")
-
-        # 添加隨機衰落 (0-5 dB)
-        fading = random.uniform(0, 5)
-        
-        # 計算 RSRP
-        rsrp = gnb.power - path_loss - fading
-        
-        return rsrp
-
-    def calculate_sinr(self, gnb, ue_position, interfering_gnbs):
-        """計算 UE 在給定位置相對於特定 gNB 的 SINR (簡化實現)
-        SINR = 接收功率 / (干擾功率 + 噪聲功率)
-        假設噪聲功率為固定值，干擾功率為其他 gNB 的接收功率之和
-        """
-        signal_power_linear = 10**(self.calculate_rsrp(gnb, ue_position) / 10.0)
-        
-        interference_power_linear = 0.0
-        for interfering_gnb in interfering_gnbs:
-            if interfering_gnb != gnb:
-                interference_power_linear += 10**(self.calculate_rsrp(interfering_gnb, ue_position) / 10.0)
-        
-        # 假設噪聲功率 (例如 -100 dBm)
-        noise_power_linear = 10**(-100 / 10.0)
-        
-        if interference_power_linear + noise_power_linear == 0:
-            return float("inf") # 避免除以零
-            
-        sinr_linear = signal_power_linear / (interference_power_linear + noise_power_linear)
-        
-        # 轉換回 dB
-        try:
-            sinr_db = 10 * math.log10(sinr_linear)
-        except ValueError:
-            return -float("inf") # 如果 sinr_linear <= 0
-            
-        return sinr_db
-
-    def is_in_coverage(self, gnb, ue_position, threshold=-110):
-        """判斷 UE 是否在 gNB 的覆蓋範圍內（基於 RSRP 閾值）"""
-        rsrp = self.calculate_rsrp(gnb, ue_position)
-        return rsrp >= threshold
-
-class ExternalChannelModelPlaceholder(ChannelModel):
-    """外部信道模型接口佔位符 (例如 Sionna)
-    實際實現需要調用外部信道模擬器的 API
-    """
-    def __init__(self, external_simulator_api=None):
-        self.api = external_simulator_api
-        print("Using External Channel Model Placeholder. Implement integration with your simulator.")
-
-    def calculate_rsrp(self, gnb, ue_position):
-        """調用外部 API 計算 RSRP"""
-        if self.api:
-            # result = self.api.get_rsrp(gnb.gnb_id, ue_position.x, ue_position.y)
-            # return result
-            pass
-        # 佔位符：返回簡化模型的結果或固定值
-        # print(f"Warning: ExternalChannelModelPlaceholder.calculate_rsrp not implemented. Using simplified model.")
-        simplified_model = SimplifiedChannelModel()
-        return simplified_model.calculate_rsrp(gnb, ue_position)
-        # raise NotImplementedError("External RSRP calculation not implemented")
-
-    def calculate_sinr(self, gnb, ue_position, interfering_gnbs):
-        """調用外部 API 計算 SINR"""
-        if self.api:
-            # interfering_ids = [ig.gnb_id for ig in interfering_gnbs if ig != gnb]
-            # result = self.api.get_sinr(gnb.gnb_id, ue_position.x, ue_position.y, interfering_ids)
-            # return result
-            pass
-        # 佔位符：返回簡化模型的結果或固定值
-        # print(f"Warning: ExternalChannelModelPlaceholder.calculate_sinr not implemented. Using simplified model.")
-        simplified_model = SimplifiedChannelModel()
-        return simplified_model.calculate_sinr(gnb, ue_position, interfering_gnbs)
-        # raise NotImplementedError("External SINR calculation not implemented")
-
-    def is_in_coverage(self, gnb, ue_position, threshold=-110):
-        """調用外部 API 判斷覆蓋範圍"""
-        if self.api:
-            # rsrp = self.api.get_rsrp(gnb.gnb_id, ue_position.x, ue_position.y)
-            # return rsrp >= threshold
-            pass
-        # 佔位符：使用簡化模型的結果
-        # print(f"Warning: ExternalChannelModelPlaceholder.is_in_coverage not implemented. Using simplified model.")
-        simplified_model = SimplifiedChannelModel()
-        return simplified_model.is_in_coverage(gnb, ue_position, threshold)
-        # raise NotImplementedError("External coverage check not implemented")
+from channel_models import (
+    Position,
+    ChannelModel,
+    SimplifiedChannelModel,
+    ExternalChannelModelPlaceholder,
+)
 
 # --- 網絡實體類 ---
